Options -xx and -sx are now supporting regular expressions
The options -xx and -sx are reading excludes from a given exclude file.
Normally the first part of the exclude pattern is name and the second
part is the type - patterns are line by line pipe delimitered.
This commit adds an option to treat the first part (name) as a regular
expression. If the line starts with '~' as character, the name can be
any regular expression. This is a simple way to e.g. exclude all sensors
with the word 'Temp' in it:
~Temp|Temperature